Complete Realtime test suite migration - Upgrade
