+++ .jshintrc
@@ -7,6 +7,7 @@
     "setInterval",
     "setTimeout",
     "clearInterval",
+    "clearTimeout",
     "global",
     "window",
     "isBrowser",
